requirements.txt
PyYAML
accelerate
datasets
deepspeed
einops
evaluate
peft
protobuf
scikit-learn
scipy
sentencepiece
torch>=2.0.1
fire
pillow
transformers
isort
black
wandb
opencv-python
jupyter
jupyterlab
matplotlib
japanize-matplotlib
